fix(Blist): The `count` change should executed in operation instead of `CreateNode()` and `FreeNode()`.
If the count is only changed in `CreateNode()` and `FreeNode()`, the
new node or extracted isn't pushed to the list is make the count be fact
count + 1, this is unexpected.
If don't change the original count change, just decrement the count in
`Extract()`, this requires the `PushFront()` and `PushBack()`(Node
overload) also increment the count.
But them are also a forwarder of `PushFront()` and `PushBack()`(Value
overload).